Bring theme to the chips and separate components from page
